Move PlainStack to own header file
